// packages/api/src/mappings/cronometer/models/diary-summary.ts
import { z } from "zod";
import { cronometerMacrosResp } from "./macros";
import { cronometerMealResp } from "./meal";

// https://cronometer.com/developer/docs/openapi.html#operation/post-diary_summary
export const cronometerDiarySummaryResp = z.object({
  completed: z.boolean(),
  day: z.string(),
  food_grams: z.number(),
  foods: z.array(cronometerMealResp).optional().nullable(),
  macros: cronometerMacrosResp,
  metrics: z.array(z.any()),
  nutrients: z.object({
    Alanine: z.number(),
    Alcohol: z.number(),
    "Alpha-carotene": z.number(),
    Arginine: z.number(),
    Ash: z.number(),
    "Aspartic acid": z.number(),
    "B1 (Thiamine)": z.number(),
    "B12 (Cobalamin)": z.number(),
    "B2 (Riboflavin)": z.number(),
    "B3 (Niacin)": z.number(),
    "B5 (Pantothenic Acid)": z.number(),
    "B6 (Pyridoxine)": z.number(),
    "Beta Tocopherol": z.number(),
    "Beta-carotene": z.number(),
    "Beta-cryptoxanthin": z.number(),
    Caffeine: z.number(),
    Calcium: z.number(),
    Carbs: z.number(),
    Cholesterol: z.number(),
    Choline: z.number(),
    Copper: z.number(),
    Cystine: z.number(),
    "Delta Tocopherol": z.number(),
    Energy: z.number(),
    Fat: z.number(),
    Fiber: z.number(),
    Folate: z.number(),
    Fructose: z.number(),
    Galactose: z.number(),
    "Gamma Tocopherol": z.number(),
    Glucose: z.number(),
    "Glutamic acid": z.number(),
    Glycine: z.number(),
    Histidine: z.number(),
    Iron: z.number(),
    Isoleucine: z.number(),
    Lactose: z.number(),
    Leucine: z.number(),
    "Lutein+Zeaxanthin": z.number(),
    Lycopene: z.number(),
    Lysine: z.number(),
    Magnesium: z.number(),
    Maltose: z.number(),
    Manganese: z.number(),
    Methionine: z.number(),
    Monounsaturated: z.number(),
    "Net Carbs": z.number(),
    Phenylalanine: z.number(),
    Phosphorus: z.number(),
    Polyunsaturated: z.number(),
    Potassium: z.number(),
    Proline: z.number(),
    Protein: z.number(),
    Retinol: z.number(),
    "Retinol Activity Equivalent": z.number().optional().nullable(),
    Saturated: z.number(),
    Selenium: z.number(),
    Serine: z.number(),
    Sodium: z.number(),
    Starch: z.number(),
    Sucrose: z.number(),
    Sugars: z.number(),
    Threonine: z.number(),
    "Trans-Fats": z.number(),
    Tryptophan: z.number(),
    Tyrosine: z.number(),
    Valine: z.number(),
    "Vitamin A": z.number(),
    "Vitamin C": z.number(),
    "Vitamin D": z.number(),
    "Vitamin E": z.number(),
    "Vitamin K": z.number(),
    Water: z.number(),
    Zinc: z.number(),
  }),
});

export type CronometerDiarySummary = z.infer<typeof cronometerDiarySummaryResp>;